requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.4.0
alembic==1.13.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.4.0
Babel==2.9.1
backports-datetime-fromisoformat==2.0.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
databases==0.9.0
ecdsa==0.19.1
environs==8.0.0
exceptiongroup==1.3.1
fastapi==0.121.3
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.7.0
packaging==25.0
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.35.0
yarl==1.22.0
